organize example files into folders and individual examples
