+++ .gitignore
@@ -1,2 +1,4 @@
 nullbytes.txt
 *.png
+
+.DS_Store
