--- examples/fake_db.txt
@@ -1,3 +1,0 @@
-{
-    "email": "dp@mayainsights.com"
-}
