+++ examples/json_response/example_json_response.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	Goster "github.com/dpouris/goster"
+)
+
+func main() {
+	g := Goster.NewServer()
+
+	g.Get("/json", func(ctx *Goster.Ctx) error {
+		response := struct {
+			Message string `json:"message"`
+		}{
+			Message: "Hello, JSON!",
+		}
+		ctx.JSON(response)
+		return nil
+	})
+
+	g.ListenAndServe(":8080")
+}
