+++ examples/middleware/example_middleware.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"log"
+
+	Goster "github.com/dpouris/goster"
+)
+
+func main() {
+	g := Goster.NewServer()
+
+	// Middleware to log requests
+	g.UseGlobal(func(ctx *Goster.Ctx) error {
+		log.Printf("Received request for %s", ctx.Request.URL.Path)
+		return nil
+	})
+
+	g.Get("/", func(ctx *Goster.Ctx) error {
+		ctx.Text("Middleware example")
+		return nil
+	})
+
+	g.ListenAndServe(":8080")
+}
